row: delete.

+++ app/controllers/rows_controller.rb
@@ -46,15 +46,16 @@ class RowsController < ApplicationController
     @row.current_count = @row.current_count - row_item.count
     @row.check_state
     order_item = row_item.order_item
-    CartItem.create count: order_item.count, total_price: order_item.price, user: current_user, product: order_item.product
-    order_item.destroy!
+    unless order_item.nil?
+      CartItem.create count: order_item.count, total_price: order_item.price, user: current_user, product: order_item.product
+      order_item.destroy!
+    end
     row_item.destroy! unless row_item.nil?
     if @row.current_count == 0
       @row.destroy!
     else
       @row.save!
     end
-    redirect_to cart_item_path
   end
 
 end
